Revise code based on feedback, add button to toggle between multiple or single ISBN scanning
