+++ openlibrary/components/BarcodeScanner.vue
@@ -1,22 +1,27 @@
 <template>
   <div id="page-barcodescanner">
-    <OLBarcodeScanner :state="state" :lastISBN="lastISBN" v-on:update:lastISBN="updateLastISBN" v-on:update:state="updateState"/>
+    <div id="interactive" class="viewport"></div>
     <div class="barcodescanner__toolbar">
       <details class="barcodescanner__advanced">
         <summary class="glass-button icon-button">
-          <svg xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:cc="http://creativecommons.org/ns#" xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:svg="http://www.w3.org/2000/svg" xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 4.2742091 4.2333331" height="16" width="16"><path d="M 1.6561242,0.00198689 C 1.5224783,0.02929715 1.3981294,0.07810697 1.2746523,0.12924106 1.2650645,0.49211882 1.3252052,0.96714289 1.1837149,1.0917824 1.0448392,1.2143878 0.58027458,1.1118292 0.22117379,1.0734786 0.14476319,1.2109015 0.08433198,1.3552972 0.03958966,1.5092805 0.31734074,1.739384 0.70229912,1.9967977 0.71159623,2.1812871 0.72089334,2.3678103 0.34755635,2.6490479 0.09421016,2.9076237 0.15435082,3.0575395 0.23947746,3.1943815 0.33041481,3.3254125 0.68399544,3.2519072 1.1244458,3.1083831 1.2746523,3.2164619 1.427764,3.3265746 1.4286355,3.8007271 1.47454,4.1606995 1.6195168,4.2042803 1.772338,4.2214213 1.9283551,4.2333334 2.0919261,3.9111305 2.2581119,3.466322 2.4367907,3.4160595 2.6201181,3.3646352 2.996651,3.6821892 3.3083946,3.8701651 3.4353582,3.7821329 3.5483762,3.6743446 3.6535496,3.5613266 3.5059581,3.2286644 3.2345987,2.8105852 3.3083946,2.6351022 c 0.073796,-0.1754829 0.566833,-0.272812 0.9079206,-0.3994851 0.00261,-0.042709 0.018304,-0.083674 0.018304,-0.1272542 0,-0.1112747 -0.020046,-0.2202252 -0.036317,-0.3268514 C 3.8461741,1.6879593 3.3458736,1.6411833 3.2540645,1.4726731 3.1631272,1.3053253 3.3946833,0.85586816 3.5082825,0.51013195 3.3914875,0.4064111 3.2645238,0.31431163 3.1268104,0.23761048 2.8356947,0.45551144 2.4966408,0.81548385 2.3095366,0.78236289 2.1253376,0.74982302 1.9242876,0.30617665 1.7284674,0.00140584 c -0.023243,0.0040672 -0.049682,-0.0046482 -0.072634,0 z M 2.0192925,1.0914918 c 0.5415565,0 0.9805543,0.4389978 0.9805543,0.9805543 0,0.5415566 -0.4389978,0.9805543 -0.9805543,0.9805543 -0.5415566,0 -0.9805543,-0.4389977 -0.9805543,-0.9805543 0,-0.5415565 0.4389977,-0.9805543 0.9805543,-0.9805543 z" style="fill:currentColor;stroke:none;"></path></svg>
+          <SettingsIcon/>
                 Advanced
         </summary>
         <div class="barcodescanner__controls">
-          <button class="barcodescanner__read-isbn glass-button">
+          <button class="barcodescanner__read-isbn glass-button" @click="handleISBNDetected" :disabled="disableISBNTextButton || canvasInactive">
             Read Text ISBN
             <br>
             <small>For books that print the ISBN without a barcode</small>
           </button>
+          <button class="barcodescanner__scan-multiple glass-button" @click="handleMultiISBN">
+            {{ readMultiISBN ? 'Read one ISBN' : 'Read multiple ISBNs'}}
+            <br>
+            <small>{{ readMultiISBN ? 'Switch to read one ISBN' : 'Switch to read multiple ISBNs'}}</small>
+          </button>
         </div>
       </details>
       <div class="barcodescanner__result-strip">
-        <LazyBookCard v-if="lastISBN" :isbn="lastISBN" :state="state"/>
+         <LazyBookCard v-for="isbnObj in isbnList" :key="isbnObj.isbn" :isbn="isbnObj.isbn" :tentativeCover="isbnObj.cover" />
         <div class="empty">
           Point your camera at a barcode! 📷
         </div>
@@ -26,67 +31,229 @@
 </template>
 
 <script>
-import OLBarcodeScanner from './BarcodeScanner/components/OLBarcodeScanner.vue';
 import LazyBookCard from './BarcodeScanner/components/LazyBookCard.vue';
-import Promise from 'promise-polyfill';
-import CONFIGS from './configs';
+import SettingsIcon from './LibraryExplorer/components/icons/SettingsIcon.vue';
+import Quagga from '@ericblade/quagga2';
+import maxBy from 'lodash/maxBy';
+import countBy from 'lodash/countBy';
+import { OCRScanner, ThrottleGrouping } from './BarcodeScanner/utils/classes.js';
 
 export default {
-    components: { OLBarcodeScanner, LazyBookCard },
+    components: { LazyBookCard, SettingsIcon },
     data() {
         return {
+            disableISBNTextButton: false,
+            targetInteractive: null,
+            readMultiISBN: location.href.includes('returnTo') ? false : true,
+            canvasInactive: true,
             lastISBN: null,
-            state: {},
+            isbnList: [],
+            seenISBN: new Set(),
+            returnTo: new URLSearchParams(location.search).get('returnTo'),
+            BOX_STYLE: {color: 'green', lineWidth: 2},
+            RESULT_BOX_STYLE: {color: 'blue', lineWidth: 2},
+            RESULT_LINE_STYLE: {color: 'red', lineWidth: 15},
+            submitISBNThrottled: new ThrottleGrouping({
+                func: this.submitISBN.bind(this),
+                // Use the most frequent
+                reducer: (groupOfArgs) => {
+                    const isbnCounts = Array.from(
+                        Object.entries(
+                            countBy(groupOfArgs, (arg) => arg[0])
+                        )
+                    );
+
+                    /* eslint-disable no-unused-vars */
+                    const mostFrequentISBN = maxBy(isbnCounts, ([isbn, count]) => count)[0];
+                    return groupOfArgs.reverse().find((args) => args[0] === mostFrequentISBN);
+                },
+                wait: 300,
+            }).asFunction(),
+            quaggaVideo: null,
+            ocrScanner: null,
         }
     },
     methods: {
-        updateLastISBN(variable) {
-            this.lastISBN = variable;
-            this.fromISBN(variable);
-        },
-
-        updateState(newState) {
-            const oldState = this.state;
-            newState = Object.assign({}, oldState, newState);
-            this.state = newState;
-        },
-
-        fromISBN(isbn) {
-            fetch(`https://${CONFIGS.OL_BASE_PUBLIC}/isbn/${isbn}.json`).then(r => r.json())
-                .then(editionRecord => {
-                    this.updateState({
-                        title: editionRecord.title,
-                        identifier: isbn,
-                        link: editionRecord.key,
-                    });
-
-                    if (editionRecord.covers) {
-                        const coverId = editionRecord.covers.find(x => x !== -1);
-                        if (coverId) {
-                            this.updateState({
-                                coverSrc: `https://covers.openlibrary.org/b/id/${coverId}-M.jpg`,
-                            });
-                        }
+        start() {
+            Quagga.init({
+                locator: {
+                    halfSample: true,
+                },
+                inputStream: {
+                    name: 'Live',
+                    type: 'LiveStream',
+                    target: this.targetInteractive,
+                    constraints: {
+                    // Vertical - This is *essential* for iPhone/iPad
+                        aspectRatio: {ideal: 720/1280},
+                    },
+                },
+                decoder: {
+                    readers: ['ean_reader']
+                },
+            }, async (err) => {
+                if (err) throw err;
+                const track = Quagga.CameraAccess.getActiveTrack();
+                if (track && typeof track.getCapabilities === 'function') {
+                    const capabilities = track.getCapabilities();
+                    // Use a higher resolution
+                    if (capabilities.width.max >= 1280 && capabilities.height.max >= 720) {
+                        await track.applyConstraints({advanced: [{width: 1280, height: 720}]});
                     }
-
-                    return fetch(`https://${CONFIGS.OL_BASE_PUBLIC}${editionRecord.works[0].key}.json`).then(r => r.json())
-                }).then(workRecord => {
-                    return Promise.all(
-                        workRecord.authors
-                            .map(a => fetch(`https://${CONFIGS.OL_BASE_PUBLIC}${a.author.key}.json`).then(r => r.json()))
-                    );
-                }).then(authorRecords => {
-                    this.updateState({
-                        loading: false,
-                        byline: authorRecords.map(a => a.name).join(', '),
-                    });
-                })
-                // eslint-disable-next-line no-unused-vars
-                .catch((err) => {
-                    this.updateState({loading: false, errored: true});
                 }
-                );
-        }
+                const quaggaVideo = this.targetInteractive.getElementsByTagName('video')[0];
+
+                const ocrScanner = new OCRScanner();
+                // document.body.append(s.userCanvas);
+                ocrScanner.onISBNDetected((isbn) => {
+                    this.submitISBNThrottled(isbn, '/static/images/openlibrary-logo-tighter.svg');
+                });
+                ocrScanner.init();
+
+                this.quaggaVideo = quaggaVideo;
+                this.ocrScanner = ocrScanner;
+                Quagga.start();
+                if (quaggaVideo.paused) {
+                    quaggaVideo.setAttribute('controls', 'true');
+                    quaggaVideo.addEventListener('play', () => quaggaVideo.removeAttribute('controls'));
+                }
+            });
+            Quagga.onProcessed(this.handleQuaggaProcessed.bind(this));
+            Quagga.onDetected(this.handleQuaggaDetected.bind(this));
+        },
+
+        async handleISBNDetected() {
+            const canvas = document.createElement('canvas');
+            canvas.width = this.quaggaVideo.videoWidth;
+            canvas.height = this.quaggaVideo.videoHeight;
+            canvas.getContext('2d').drawImage(this.quaggaVideo, 0, 0, canvas.width, canvas.height);
+
+            this.cameraFlash(this.quaggaVideo);
+
+            this.disableISBNTextButton = true;
+            await this.ocrScanner.init();
+            await this.ocrScanner.doOCR(canvas);
+            this.disableISBNTextButton = false;
+        },
+
+        async cameraFlash(el) {
+            el.style.animation = 'camera-flash 0.2s';
+            await Promise.race([
+                new Promise((res) => setTimeout(res, 2000)),
+                new Promise((res) => el.addEventListener('animationend', res, {once: true})),
+            ]);
+            el.style.animation = '';
+        },
+
+        handleQuaggaProcessed(result) {
+            if (!result) return;
+
+            const drawingCtx = Quagga.canvas.ctx.overlay;
+
+            if (result.boxes) {
+                this.clearOverlays();
+
+                result.boxes.forEach(box => {
+                    if (box !== result.box) {
+                        Quagga.ImageDebug.drawPath(box, {x: 0, y: 1}, drawingCtx, this.BOX_STYLE);
+                    }
+                });
+            }
+
+            if (result.box) {
+                Quagga.ImageDebug.drawPath(result.box, {x: 0, y: 1}, drawingCtx, this.RESULT_BOX_STYLE);
+                this.clearLater();
+            }
+
+            if (result.codeResult && result.codeResult.code && this.isBarcodeISBN(result.codeResult.code)) {
+                Quagga.ImageDebug.drawPath(result.line, {x: 'x', y: 'y'}, drawingCtx, this.RESULT_LINE_STYLE);
+                this.clearLater();
+            }
+        },
+
+        clearLater() {
+            if (this.clearTimeout) clearTimeout(this.clearTimeout);
+            this.clearTimeout = setTimeout(this.clearOverlays, 500);
+        },
+
+        clearOverlays() {
+            if (this.clearTimeout) clearTimeout(this.clearTimeout);
+            const drawingCtx = Quagga.canvas.ctx.overlay;
+            const drawingCanvas = Quagga.canvas.dom.overlay;
+            const canvasWidth = parseFloat(drawingCanvas.getAttribute('width'));
+            const canvasHeight = parseFloat(drawingCanvas.getAttribute('height'));
+            drawingCtx.clearRect(0, 0, canvasWidth, canvasHeight);
+        },
+
+        handleQuaggaDetected(result) {
+            const code = result.codeResult.code;
+            if (!this.isBarcodeISBN(code)) return;
+
+            this.submitISBNThrottled(code, Quagga.canvas.dom.image.toDataURL());
+        },
+
+        submitISBN(isbn, tentativeCoverUrl) {
+            if (isbn === this.lastISBN) return;
+            if (this.seenISBN.has(isbn)) return;
+
+            if (this.returnTo) {
+                location = this.returnTo.replace('$$$', isbn);
+            }
+            this.isbnList.unshift({isbn: isbn, cover: tentativeCoverUrl});
+            this.seenISBN.add(isbn);
+        },
+
+        isBarcodeISBN(code) {
+            return code.startsWith('97');
+        },
+
+        handleMultiISBN() {
+            if (location.href.includes('returnTo')) {
+                window.history.replaceState('', '', 'barcodescanner');
+                this.readMultiISBN = true;
+                this.returnTo = '';
+            } else {
+                window.history.replaceState('', '', 'barcodescanner?returnTo=/isbn/$$$');
+                this.readMultiISBN = false;
+                this.returnTo = new URLSearchParams(location.search).get('returnTo');
+            }
+        },
+
+        waitForShadowDOM() {
+            return new Promise(resolve => {
+                if (document.querySelector('#interactive')) {
+                    return resolve(document.querySelector('#interactive'));
+                }
+                if (document.querySelector('ol-barcode-scanner').shadowRoot.getElementById('interactive')) {
+                    return resolve(document.querySelector('ol-barcode-scanner').shadowRoot.getElementById('interactive'))
+                }
+                const observer = new MutationObserver(_mutations => {
+                    if (document.querySelector('ol-barcode-scanner').shadowRoot.getElementById('interactive')) {
+                        observer.disconnect();
+                        resolve(document.querySelector('ol-barcode-scanner').shadowRoot.getElementById('interactive'));
+                    }
+                })
+                observer.observe(document.body, {
+                    childList: true,
+                    subtree: true,
+                })
+            })
+        },
+    },
+    async mounted() {
+        this.targetInteractive = await this.waitForShadowDOM();
+        this.start();
+        // Disable Read Text ISBN button until camera permissions are granted
+        const observer = new MutationObserver(_mutations => {
+            if (document.querySelector('ol-barcode-scanner').shadowRoot.querySelector('.drawingBuffer')) {
+                this.canvasInactive = false;
+                observer.disconnect();
+            }
+        });
+        observer.observe(document.querySelector('ol-barcode-scanner').shadowRoot, {
+            childList: true,
+            subtree: true
+        });
     }
 }
 </script>
@@ -251,72 +418,6 @@ export default {
       align-items: center;
       flex-shrink: 0;
       border-radius: 4px;
-    }
-  }
-  a.lazy-book-card {
-    background: hsl(0, 0%, 100%);
-    flex-shrink: 0;
-    margin-right: 6px;
-    width: 80vw;
-    max-width: 300px;
-    height: 120px;
-    border: 1px solid hsl(0, 0%, 67%);
-    border-radius: 4px;
-    display: flex;
-    overflow: hidden;
-    color: hsl(0, 0%, 0%);
-    text-decoration: none;
-    position: relative;
-    transition-property: background-color border-color;
-    transition-duration: .2s;
-
-    &:first-child { animation: slideUp .8s; }
-    &:nth-child(2) { animation: shiftRight .8s; }
-    &:hover {
-      background: lighten(hsl(240, 100%, 50%), 45%);
-      border-color: lighten(hsl(240, 100%, 50%), 15%);
-      color: lighten(hsl(240, 100%, 50%), 10%);
-    }
-    &::before {
-      display: none;
-      content: "";
-      position: absolute;
-      width: 6px;
-      height: 6px;
-      border-radius: 100px;
-      margin: 10px;
-      right: 0;
-    }
-    &.loading::before {
-      display: block;
-      background: hsl(240, 100%, 50%);
-      opacity: 0;
-      animation: pulse .5s infinite alternate;
-      // Only show loading animation if it takes
-      // longer than this
-      animation-delay: .5s;
-    }
-    &.errored::before {
-      display: block;
-      background: hsl(8, 78%, 49%);
-    }
-    .title { font-weight: bold; line-height: 1.1em; }
-    .byline { font-size: .9em; margin-top: .1em; }
-    .cover {
-      width: 25%;
-      img { width: 100%; height: 100%; object-fit: cover; }
-    }
-    .identifier {
-      margin-top: 4px;
-      padding-top: 4px;
-      color: hsl(0, 0%, 33%);
-      border-top: 1px dotted;
-      font-family: monospace;
-      font-size: 1.2em;
-    }
-    .info {
-      flex: 1;
-      padding: 8px;
     }
   }
 </style>
